requirements.txt
psycopg2
gunicorn
anyio==3.3.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Django==3.2.9
django-ninja==0.16.1
django-ninja-apikey==0.1.0
fake-useragent==0.1.11
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.3
httpx==0.21.1
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
lxml==4.6.4
orjson==3.6.4
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.0